requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
alembic
pydantic-settings
python-jose[cryptography]
python-multipart
bcrypt==4.0.1